Design update: moved saved search under securitydata module
