+++ src/py42/_internal/clients/securitydata.py
@@ -6,3 +6,66 @@ class SecurityClient(BaseClient):
         uri = u"/c42api/v3/SecurityEventsLocation"
         params = {u"userUid": user_uid}
         return self._session.get(uri, params=params)
+
+
+class SavedSearchClient(BaseClient):
+    """A client to interact with saved search APIs."""
+
+    _version = u"v1"
+    _resource = u"/forensic-search/queryservice/api/{}/saved".format(_version)
+
+    _response_header = [u"name", u"id", u"notes", u"modifiedByUsername", u"modifiedTimestamp"]
+
+    def __init__(self, session, file_event_client):
+        super(SavedSearchClient, self).__init__(session)
+        self._file_event_client = file_event_client
+
+    def _parse_response(self, response):
+        if u"searches" in response:
+            return {
+                key: search[key]
+                for search in response[u"searches"]
+                for key in self._response_header
+            }
+
+    def get(self):
+        """Fetch details of existing saved searches.
+        Provides details such as saved search id, name, notes, modifier name, last modified time.
+
+        Returns:
+            dict
+        """
+        uri = u"{}".format(self._resource)
+        response = self._session.get(uri)
+        return self._parse_response(response)
+
+    def get_by_id(self, search_id):
+        """Fetch details of saved search for given search Id.
+
+        Args:
+            search_id (str): Unique search Id of the saved search, Required.
+        Returns:
+            :class:`py42.response.Py42Response`
+        """
+        uri = u"{}/{}".format(self._resource, search_id)
+        return self._session.get(uri)
+
+    @staticmethod
+    def _make_query(response):
+        query = dict(pgNum=1, pgSize=100, purpose=u"USER_EXECUTED_SEARCH")
+        query["groups"] = response[u"searches"][0][u"groups"]
+        return query
+
+    def execute(self, search_id):
+        """
+        Execute a saved search for given search Id and return incremental results based
+        on the previous execution.
+
+        Args:
+            search_id (str): Unique search Id of the saved search, Required.
+        Returns:
+            :class:`py42.response.Py42Response`
+        """
+        response = self.get_by_id(search_id)
+        query = SavedSearchClient._make_query(response)
+        return self._file_event_client.search(query=query)
