+++ src/py42/modules/securitydata.py
@@ -14,6 +14,15 @@ class SecurityModule(object):
         self._client_cache = {}
         self._client_cache_lock = Lock()
 
+    @property
+    def savedsearches(self):
+        """A collection of methods related to retrieving forensic search data.
+
+        Returns:
+            :class: `py42.clients.investigation.SavedSearchClient`
+        """
+        return self._microservices_client_factory.get_saved_search_client()
+
     def get_security_plan_storage_info_list(self, user_uid):
         """Gets IDs (plan UID, node GUID, and destination GUID) for the storage nodes containing
         the file activity event data for the user with the given UID.
